refactor: move anr code
