+++ embrace-android-instrumentation-anr/src/main/kotlin/io/embrace/android/embracesdk/internal/instrumentation/anr/AnrServiceSupplier.kt
@@ -2,7 +2,7 @@ package io.embrace.android.embracesdk.internal.instrumentation.anr
 
 import android.os.Looper
 import io.embrace.android.embracesdk.internal.arch.InstrumentationArgs
-import io.embrace.android.embracesdk.internal.instrumentation.anr.detection.BlockedThreadDetector
+import io.embrace.android.embracesdk.internal.instrumentation.thread.blockage.BlockedThreadDetector
 import io.embrace.android.embracesdk.internal.worker.Worker
 
 /**
