+++ embrace-android-instrumentation-anr/src/test/kotlin/io/embrace/android/embracesdk/fakes/FakeThreadBlockageListener.kt
@@ -1,7 +1,7 @@
 package io.embrace.android.embracesdk.fakes
 
-import io.embrace.android.embracesdk.internal.instrumentation.anr.detection.ThreadBlockageEvent
-import io.embrace.android.embracesdk.internal.instrumentation.anr.detection.ThreadBlockageListener
+import io.embrace.android.embracesdk.internal.instrumentation.thread.blockage.ThreadBlockageEvent
+import io.embrace.android.embracesdk.internal.instrumentation.thread.blockage.ThreadBlockageListener
 
 internal class FakeThreadBlockageListener : ThreadBlockageListener {
     var intervalCount: Int = 0
